requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiomysql==0.2.0
cryptography==41.0.7
alembic==1.13.0